Add files via upload
On Christmas Day, 1994, a hacker launched a sophisticated IP spoofing attack against Tsutomu Shimomura’s computers in San Diego SupercomputerCenter”This hacker turned out to be Kevin Mitnick. 
This repo is a recreation of the IP spoofing attack by exploiting the fact that ACK numbers can be predicted and exploiting the trust relationship between x-terminal, and server.